Relocate CSV tests to their own file, add comment
