+++ tests/app/public_test.py
@@ -30,17 +30,6 @@ def test_get_state_info(app):
     assert respjson[0]["pum"] == False
     assert respjson[0]["notes"] == 'Testing123'
 
-    resp = client.get("/api/v1/public/states/info.csv")
-    assert resp.status_code == 200
-    lines = resp.data.decode("utf-8").splitlines()
-    reader = csv.DictReader(lines, delimiter=',')
-    cnt = 0
-    for row in reader:
-        cnt += 1
-        assert row["State"] in ["NY", "WA"]
-        assert row["Notes"] in ["Testing123", "Testing321"]
-    assert cnt == 2
-
 
 def test_get_states_daily(app, headers):
     # post some test data
@@ -97,16 +86,6 @@ def test_get_states_daily(app, headers):
     assert resp.status_code == 200
     assert resp.json == []
 
-    resp = client.get("/api/v1/public/states/daily.csv")
-    assert resp.status_code == 200
-    lines = resp.data.decode("utf-8").splitlines()
-    assert len(lines) == 57
-    reader = csv.DictReader(lines, delimiter=',')
-    data = list(reader)
-    assert data[0]["Date"] == '20200618'
-    assert data[0]["Positive"] == "708"
-    assert data[0]["Negative"] == "80477"
-
 
 def test_get_us_daily_column_names(app):
     colnames = CoreData.numeric_fields()
@@ -154,83 +133,6 @@ def test_get_us_daily(app, headers):
     assert resp.json[1]['date'] == '2020-05-24'
     assert resp.json[1]['positive'] == 24
     assert resp.json[1]['negative'] == 12
-
-
-def test_get_us_states_csv(app, headers):
-    test_data = daily_push_ny_wa_two_days()
-    client = app.test_client()
-
-    # Write a batch containing the above data, two days each of NY and WA
-    resp = client.post(
-        "/api/v1/batches",
-        data=json.dumps(test_data),
-        content_type='application/json',
-        headers=headers)
-    assert resp.status_code == 201
-    batch_id = resp.json['batch']['batchId']
-
-    # Publish the new batch
-    resp = client.post("/api/v1/batches/{}/publish".format(batch_id),
-                       headers=headers)
-    assert resp.status_code == 201
-
-    # test US daily CSV
-    resp = client.get("/api/v1/public/us/daily.csv")
-    assert resp.status_code == 200
-    lines = resp.data.decode("utf-8").splitlines()
-    assert len(lines) == 3
-    reader = csv.DictReader(lines, delimiter=',')
-    data = list(reader)
-
-    assert data[0]["Date"] == '20200525'
-    assert data[0]["States"] == "2"
-    assert data[0]["Positive"] == "30"
-    assert data[0]["Negative"] == "15"
-
-    assert data[1]["Date"] == '20200524'
-    assert data[1]["States"] == "2"
-    assert data[1]["Positive"] == "24"
-    assert data[1]["Negative"] == "12"
-
-    # test states current CSV
-    resp = client.get("/api/v1/public/states/current.csv")
-    assert resp.status_code == 200
-    lines = resp.data.decode("utf-8").splitlines()
-    assert len(lines) == 3
-    reader = csv.DictReader(lines, delimiter=',')
-    data = list(reader)
-
-    # test states current CSV
-    resp = client.get("/api/v1/public/states/current.csv")
-    assert resp.status_code == 200
-    lines = resp.data.decode("utf-8").splitlines()
-    assert len(lines) == 3
-    reader = csv.DictReader(lines, delimiter=',')
-    data = list(reader)
-
-    assert len(data) == 2
-    assert "Date" not in data[0]
-    assert data[0]["State"] == "NY"
-    assert data[0]["Positive"] == "20"
-    assert data[0]["Negative"] == "5"
-    assert data[1]["State"] == "WA"
-    assert data[1]["Positive"] == "10"
-    assert data[1]["Negative"] == "10"
-
-    # test US current CSV
-    resp = client.get("/api/v1/public/us/current.csv")
-    assert resp.status_code == 200
-    lines = resp.data.decode("utf-8").splitlines()
-    assert len(lines) == 2
-    reader = csv.DictReader(lines, delimiter=',')
-    data = list(reader)
-
-    # US current shouldn't have date and state columns
-    assert "Date" not in data[0]
-    assert "State" not in data[0]
-    assert "Deaths" in data[0]
-    assert data[0]["Positive"] == "30"
-    assert data[0]["Negative"] == "15"
 
 
 def test_get_states_daily_for_state(app, headers):
